+++ riverdriver/riversqlite/internal/dbsqlc/river_queue.sql.go
@@ -63,6 +63,7 @@ WHERE name IN (
     SELECT name
     FROM /* TEMPLATE: schema */river_queue
     WHERE river_queue.updated_at < ?1
+        AND (/* TEMPLATE_BEGIN: queues_included_clause */ true /* TEMPLATE_END */)
     ORDER BY name ASC
     LIMIT ?2
 )